Added ability to save and load locations using Realm. This is a mobile database that runs on SQLite and works really well. You may need to follow this step to get Realm working: "In the Android Studio menu: Tools->Android->Sync Project with Gradle Files"
Added a bar that appears whenever you actually add a marker, but we nee to figure out how to actually grab the marker from the ID we have stored. I commented the line where this needs to be done. Once we can figure that out all the rest of the steps should be pretty easy.